完成贴图加载，实现图形绘制方式切换，修正圆形与矩形在不同绘制方式下的节点表现，

+++ cg-1/tools/creator.cpp
@@ -17,7 +17,7 @@ void RubberBandCreator::dragMove(vec3 end) {
 void LineCreator::dragStart(vec3 start, vec3 end) {
     result = new Line(start, end);
     result->setColor(mainScene->currentColor);
-    result->setZIndex(mainScene->shapes.size());
+    result->setZIndex((int) mainScene->shapes.size());
     mainScene->shapes.push_back(result);
 }
 
@@ -25,20 +25,20 @@ void TriangleCreator::dragStart(vec3 start, vec3 end) {
     result = new Triangle(start, end);
     result->setColor(mainScene->currentColor);
     result->setMode(mainScene->currentDrawMode);
-    result->setZIndex(mainScene->shapes.size());
+    result->setZIndex((int) mainScene->shapes.size());
     mainScene->shapes.push_back(result);
 }
 void RectCreator::dragStart(vec3 start, vec3 end) {
     result = new Rect(start, end);
     result->setColor(mainScene->currentColor);
     result->setMode(mainScene->currentDrawMode);
-    result->setZIndex(mainScene->shapes.size());
+    result->setZIndex((int) mainScene->shapes.size());
     mainScene->shapes.push_back(result);
 }
 void CircleCreator::dragStart(vec3 start, vec3 end) {
     result = new Circle(start, end);
     result->setColor(mainScene->currentColor);
     result->setMode(mainScene->currentDrawMode);
-    result->setZIndex(mainScene->shapes.size());
+    result->setZIndex((int) mainScene->shapes.size());
     mainScene->shapes.push_back(result);
 }
